% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copilot.R
\name{wkcopilot_command}
\alias{wkcopilot_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkcopilot_command(copilot, cmd_json)
}
\arguments{
\item{copilot}{A copilot handle from [wkcopilot_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
